pyproject.toml
[project]
name = "signature-rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.8.1",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.2",
    "colorama>=0.4.6",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "dnspython>=2.7.0",
    "dotenv>=0.9.9",
    "idna>=3.10",
    "iniconfig>=2.1.0",
    "mando>=0.8.2",
    "numpy>=2.3.0",
    "packaging>=25.0",
    "pandas>=2.3.0",
    "pluggy>=1.6.0",
    "py3-validate-email>=1.0.5.post2",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "radon>=1.3.1",
    "requests>=2.32.4",
    "shortuuid>=1.0.13",
    "six>=1.17.0",
    "sqlparse>=0.5.3",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "gunicorn>=23.0.0",
]
